LayoutTests/svg/dynamic-updates/SVGFilterElement-dom-filterRes-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'filterRes' attribute of the SVGFilterElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS filterElement.getAttribute('filterRes') is "400"
PASS filterElement.getAttribute('filterRes') is "200"
PASS successfullyParsed is true

TEST COMPLETE
